requirements.txt
hydra-core
librosa
soundfile
numpy==1.23.5
pytorch-lightning